requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
cloudinary==1.29.0
Django==4.0.4
django-cloudinary-storage==0.3.0
django-cors-headers==3.11.0
django-jazzmin==2.5.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
Pillow==9.1.0
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0
